+++ lib/src/synthesizers/systemverilog/systemverilog.dart
@@ -1,6 +1,7 @@
-// Copyright (C) 2021-2024 Intel Corporation
+// Copyright (C) 2021-2026 Intel Corporation
 // SPDX-License-Identifier: BSD-3-Clause
 
 export 'sv_service.dart';
 export 'systemverilog_mixins.dart';
 export 'systemverilog_synthesizer.dart';
+export 'systemverilog_synthesizer_configuration.dart';