+++ rohd_devtools_extension/lib/rohd_devtools/cubit/signal_search_term_cubit.dart
@@ -1,4 +1,4 @@
-// Copyright (C) 2025 Intel Corporation
+// Copyright (C) 2025-2026 Intel Corporation
 // SPDX-License-Identifier: BSD-3-Clause
 //
 // signal_search_term_cubit.dart
@@ -9,9 +9,12 @@
 
 import 'package:flutter_bloc/flutter_bloc.dart';
 
+/// Cubit that stores the current signal-table search term.
 class SignalSearchTermCubit extends Cubit<String?> {
+  /// Creates the signal-search cubit with no initial term.
   SignalSearchTermCubit() : super(null);
 
+  /// Updates the search term.
   void setTerm(String term) {
     emit(term);
   }